requirements.txt
Flask==3.0.3
streamlit==1.37.1
pandas==2.2.2
openai==0.28.0
spacy==3.7.5
pyyaml==6.0.2